feat: move input to ui/input-fields components
