+++ project/components/modals/edit-project-modal.tsx
@@ -35,9 +35,9 @@ import { updateProject } from "@/lib/actions/projects";
 import { ProjectSchema } from "@/types/projects";
 import { redirect, RedirectType } from "next/navigation";
 import { useEffect, useState } from "react";
-import Input from "../input";
 import Textarea from "../textarea";
 import Button from "../ui/buttons/button";
+import Input from "../ui/input-fields/input";
 import { showErrorToast, showSuccessToast } from "../ui/toast";
 import Modal from "./modal";
