+++ project/components/sections/settings/profile-settings-form.tsx
@@ -1,6 +1,6 @@
-import Input from "@/components/input";
 import Select from "@/components/select";
 import Button from "@/components/ui/buttons/button";
+import Input from "@/components/ui/input-fields/input";
 import { cn } from "@/lib/utils";
 
 interface ProfileSettingsFormProps {
